requirements.txt
discord
discord-py-slash-command
python-dotenv
scrapy
scrapy-user-agents
